working with production directory Signed-off-by:Matthew Scaperoth <mscapero@gwu.edu>

+++ protected/controllers/ApiController.php
@@ -393,7 +393,7 @@ class ApiController extends CController {
             }
             if (file_exists($target))
                 rmdir($target);
-        } elseif (is_file($target)) {
+        } elseif (is_file($target)&& $target->getFilename() != 'Thumbs.db') {
             unlink($target);
         }
     }
